+++ base/options.jl
@@ -24,6 +24,7 @@ struct JLOptions
     startupfile::Int8
     compile_enabled::Int8
     code_coverage::Int8
+    code_coverage_mode::Int8
     malloc_log::Int8
     tracked_path::Ptr{UInt8}
     opt_level::Int8